Revert accidental commit of "%z" support in r26071 (the implementation assumed size_t==long, which is not always the case in hosted environments) Remove four remaining uses of %z, three of which were in DEBUGF
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@27479 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/mpegplayer/alloc.c
@@ -96,8 +96,8 @@ static void * mpeg_malloc_internal (unsigned char *mallocbuf,
 {
     void *x;
 
-    DEBUGF("mpeg_alloc_internal: bs:%zu s:%u reason:%s (%d)\n",
-           bufsize, size, mpeg_get_reason_str(reason), reason);
+    DEBUGF("mpeg_alloc_internal: bs:%lu s:%u reason:%s (%d)\n",
+           (unsigned long)bufsize, size, mpeg_get_reason_str(reason), reason);
 
     if ((size_t) (*mem_ptr + size) > bufsize)
     {
